update tests, new shared_functions.js
