+++ test/payout_game_session.js
@@ -1,3 +1,5 @@
+import sharedFunctions from './shared_functions';
+
 const JoyToken = artifacts.require('JoyToken');
 const JoyTokenUpgraded = artifacts.require('JoyTokenUpgraded');
 const GameDeposit = artifacts.require('GameDeposit');
@@ -5,6 +7,7 @@ const JoyGamePlatform = artifacts.require('JoyGamePlatform');
 const Web3 = require('web3');
 
 contract('PayOut_Game_Result', (accounts) => {
+	const { payoutAndCheck } = sharedFunctions;
 	const web3 = new Web3();
 	const { BN } = web3.utils;
 	web3.setProvider(JoyTokenUpgraded.web3.currentProvider);
@@ -54,66 +57,19 @@ contract('PayOut_Game_Result', (accounts) => {
 	afterEach(async () => {
 		const testGameHash = web3.utils.randomHex(32);
 
+		// unlock all
+		const remainBalance = '0';
+
 		const playerLockedBalance = await depositInstance.playerLockedFunds(testPlayer, joyGameInstance.address);
 		if (playerLockedBalance.gt(new BN('0'))) {
 			await joyGameInstance.payOutGameResult(
 				testPlayer,
+				remainBalance,
 				playerLockedBalance,
-				testGameHash,
-				{ from: platformOwner }
+				testGameHash, { from: platformOwner }
 			);
 		}
 	});
-
-	function checkAccountedBalances(finalBalance, reserveWinnings, gameDevWinnings) {
-		// generate randon hex bytes32 as gameHash
-		const testGameHash = web3.utils.randomHex(32);
-
-		return new Promise(async (resolve, reject) => {
-			const initialPlayerBalance = await joyTokenERC223.balanceOf(testPlayer);
-			const initialReserveBalance = await depositInstance.balanceOfPlayer(platformReserve);
-			const initialGameDevBalance = await depositInstance.balanceOfPlayer(gameDeveloper);
-
-			try {
-				await joyGameInstance.payOutGameResult(
-					testPlayer,
-					finalBalance.toString(),
-					testGameHash,
-					{ from: platformOwner }
-				);
-			} catch (err) {
-				reject(err);
-				return;
-			}
-
-			// balance in wallet
-			const playerBalance = await joyTokenERC223.balanceOf(testPlayer);
-
-			assert.ok(
-				playerBalance.eq(initialPlayerBalance.add(finalBalance)),
-				'Player balance should by equal finalBalance.'
-			);
-
-			const lockedFunds = await depositInstance.playerLockedFunds(testPlayer, joyGameInstance.address);
-			assert.ok(
-				lockedFunds.eq(new BN('0')),
-				'Player locked funds should by equal to 0.'
-			);
-
-			const reserveBalance = await depositInstance.balanceOfPlayer(platformReserve);
-			assert.ok(
-				reserveBalance.eq(initialReserveBalance.add(reserveWinnings)),
-				'Reserve balance should be inceased.'
-			);
-
-			const gameDevBalance = await depositInstance.balanceOfPlayer(gameDeveloper);
-			assert.ok(
-				gameDevBalance.eq(initialGameDevBalance.add(gameDevWinnings)),
-				'GameDev balance should be inceased.'
-			);
-			resolve();
-		});
-	}
 
 	it('player_win_unsuccessful', (done) => {
 		const finalBalance = (new BN(testAmount)).mul(new BN('2'));
@@ -121,7 +77,7 @@ contract('PayOut_Game_Result', (accounts) => {
 		const reserveWinnings = (new BN(testAmount)).neg();
 		const gameDevWinnings = new BN('0');
 
-		checkAccountedBalances(finalBalance, reserveWinnings, gameDevWinnings)
+		payoutAndCheck(testPlayer, finalBalance, reserveWinnings, gameDevWinnings)
 			.catch((err) => {
 				assert.include(
 					err.message,
@@ -144,7 +100,7 @@ contract('PayOut_Game_Result', (accounts) => {
 			joyTokenInstance.approve(joyTokenERC223.address, testAmount, { from: platformReserve }),
 			joyTokenERC223.transfer(depositInstance.address, testAmount, { from: platformReserve })
 		])
-			.then(() => checkAccountedBalances(finalBalance, reserveWinnings, gameDevWinnings))
+			.then(() => payoutAndCheck(testPlayer, finalBalance, reserveWinnings, gameDevWinnings))
 			.then(() => done());
 	});
 
@@ -154,7 +110,7 @@ contract('PayOut_Game_Result', (accounts) => {
 		const reserveWinnings = (new BN(testAmount)).div(new BN('2'));
 		const gameDevWinnings = reserveWinnings;
 
-		checkAccountedBalances(finalBalance, reserveWinnings, gameDevWinnings)
+		payoutAndCheck(testPlayer, finalBalance, reserveWinnings, gameDevWinnings)
 			.then(() => done());
 	});
 
@@ -164,7 +120,7 @@ contract('PayOut_Game_Result', (accounts) => {
 		const reserveWinnings = (new BN(testAmount)).div(new BN('2'));
 		const gameDevWinnings = reserveWinnings.sub(new BN('1'));
 
-		checkAccountedBalances(finalBalance, reserveWinnings, gameDevWinnings)
+		payoutAndCheck(testPlayer, finalBalance, reserveWinnings, gameDevWinnings)
 			.then(() => done());
 	});
 
@@ -173,7 +129,7 @@ contract('PayOut_Game_Result', (accounts) => {
 		const reserveWinnings = finalBalance.div(new BN('2'));
 		const gameDevWinnings = finalBalance.div(new BN('2'));
 
-		checkAccountedBalances(finalBalance, reserveWinnings, gameDevWinnings)
+		payoutAndCheck(testPlayer, finalBalance, reserveWinnings, gameDevWinnings)
 			.then(() => done());
 	});
 
@@ -182,7 +138,7 @@ contract('PayOut_Game_Result', (accounts) => {
 		const reserveWinnings = new BN('0');
 		const gameDevWinnings = reserveWinnings;
 
-		checkAccountedBalances(finalBalance, reserveWinnings, gameDevWinnings)
+		payoutAndCheck(testPlayer, finalBalance, reserveWinnings, gameDevWinnings)
 			.then(() => done());
 	});
 });
