+++ test/transfer_to_game.js
@@ -43,9 +43,15 @@ contract('JoyToken_TransferToGame', (accounts) => {
 		const testGameHash = web3.utils.randomHex(32);
 		// playes lose all
 		const finalBalance = '0';
+		const remainBalance = '0';
 		const platformOwner = await depositInstance.owner();
 
-		await joyGameInstance.accountGameResult(accounts[5], finalBalance, testGameHash, { from: platformOwner });
+		await joyGameInstance.accountGameResult(
+			accounts[5],
+			remainBalance,
+			finalBalance,
+			testGameHash, { from: platformOwner }
+		);
 	});
